Implement feature system and Hologram API
On its own, this does not do anything, but allows for third party plugins (mainly arena extensions) to have access to APIs that act more as a service provider.

As an example, the Hologram API offered in this commit lets third party plugins interface with a common API that currently creates holograms if *either* FancyHolograms or DecentHolograms is installed. This can realistically be expanded to work with any hologram plugin, but still use the same API.

This prevents developers from needing to implement support for multiple hologram plugins in their arena extensions, and rather just offers a common system they can use instead. This may further be expanded in the future to other features, like NPCs.