Add a BattleArenaApi class
This is intended to be a stable entrypoint into BattleArena. While most of the plugin is designed to be safe regardless of whether it is "API" or not, this is a safer option for third parties in the event major changes need to occur internally.

This also allows a slight bit more flexibility with what exists in the main BattleArena class, as its functionality should be depended on less by implementing plugins. The former `BattleArena.getInstance()` entrypoint is *not* deprecated and offers a handful more methods, but should only be used in events where that is needed.